🚚 Rename: HeroBanner, MainSectionLayout 컴포넌트 (main)/ui로 위치 이동
