+++ src/app/(public)/(explore)/(main)/ui/MainSectionLayout.tsx
